Move to lazy-loaded component.
